refactor: move the build branding method inside the service
